Move repo metadata endpoint to preference
